+++ __tests__/__fixtures__/basic/not-affected.js
@@ -1,1 +1,1 @@
-// not affected
+require('./not-modified')